Feedback Cycle 5: DRYing a bit more
