+++ Features/StatusBarView.swift
@@ -15,27 +15,17 @@ struct StatusBarView: View {
             Group {
                 // Options for third split
                 HStack {
-                    Button(action: { windowManager.resize(.third(.left)) }) {
-                        Label("Left 1/3", systemImage: "rectangle.lefthalf.filled")
-                    }
-                    Button(action: { windowManager.resize(.third(.center)) }) {
-                        Label("Center 1/3", systemImage: "rectangle.center.filled")
-                    }
-                    Button(action: { windowManager.resize(.third(.right)) }) {
-                        Label("Right 1/3", systemImage: "rectangle.righthalf.filled")
-                    }
+                    button(for: .third(.left))
+                    button(for: .third(.center))
+                    button(for: .third(.right))
                 }
 
                 Divider()
 
                 // Options for half split
                 HStack {
-                    Button(action: { windowManager.resize(.half(.left)) }) {
-                        Label("Left Half", systemImage: "rectangle.lefthalf.filled")
-                    }
-                    Button(action: { windowManager.resize(.half(.right)) }) {
-                        Label("Right Half", systemImage: "rectangle.righthalf.filled")
-                    }
+                    button(for: .half(.left))
+                    button(for: .half(.right))
                 }
 
                 Divider()
@@ -43,20 +33,12 @@ struct StatusBarView: View {
                 // Corner options
                 Grid {
                     GridRow {
-                        Button(action: { windowManager.resize(.corner(.topLeft)) }) {
-                            Label("Top Left", systemImage: "arrow.up.left.square")
-                        }
-                        Button(action: { windowManager.resize(.corner(.topRight)) }) {
-                            Label("Top Right", systemImage: "arrow.up.right.square")
-                        }
+                        button(for: .corner(.topLeft))
+                        button(for: .corner(.topRight))
                     }
                     GridRow {
-                        Button(action: { windowManager.resize(.corner(.bottomLeft)) }) {
-                            Label("Bottom Left", systemImage: "arrow.down.left.square")
-                        }
-                        Button(action: { windowManager.resize(.corner(.bottomRight)) }) {
-                            Label("Bottom Right", systemImage: "arrow.down.right.square")
-                        }
+                        button(for: .corner(.bottomLeft))
+                        button(for: .corner(.bottomRight))
                     }
                 }
 
@@ -64,21 +46,12 @@ struct StatusBarView: View {
 
                 // Other options
                 HStack {
-                    Button(action: { windowManager.resize(.vertical(.top)) }) {
-                        Label("Top Half", systemImage: "rectangle.tophalf.filled")
-                    }
-                    Button(action: { windowManager.resize(.vertical(.bottom)) }) {
-                        Label("Bottom Half", systemImage: "rectangle.bottomhalf.filled")
-                    }
+                    button(for: .vertical(.top))
+                    button(for: .vertical(.bottom))
                 }
 
-                Button(action: { windowManager.resize(.generic(.fullScreen)) }) {
-                    Label("Full Screen", systemImage: "rectangle.fill")
-                }
-
-                Button(action: { windowManager.resize(.generic(.center)) }) {
-                    Label("Center", systemImage: "rectangle.center.filled")
-                }
+                button(for: .generic(.fullScreen))
+                button(for: .generic(.center))
             }
             .buttonStyle(.bordered)
 
@@ -89,5 +62,16 @@ struct StatusBarView: View {
             }
         }
         .padding()
+    }
+}
+
+private extension StatusBarView {
+    func button(for screenPosition: ScreenPosition) -> some View {
+        Button {
+            windowManager.resize(screenPosition)
+        } label: {
+            let configuration = ModernButtonConfig.config(for: screenPosition)
+            Label(configuration.title, systemImage: configuration.systemIconName)
+        }
     }
 }
