+++ Utils/KeyboardShortcuts.swift
@@ -131,55 +131,3 @@ class HotKeyCallbacks {
 
     private init() {}
 }
-
-// AppDelegate.swift - Install an event handler
-import Cocoa
-
-class AppDelegate: NSObject, NSApplicationDelegate {
-    func applicationDidFinishLaunching(_: Notification) {
-        // Install the event handler
-        installEventHandler()
-
-        // Initialize keyboard shortcuts
-        _ = KeyboardShortcuts.shared
-    }
-
-    private func installEventHandler() {
-        // Define the event type for hot key pressed events
-        var eventType = EventTypeSpec(
-            eventClass: OSType(kEventClassKeyboard),
-            eventKind: UInt32(kEventHotKeyPressed)
-        )
-
-        // Install the event handler
-        InstallEventHandler(
-            GetEventDispatcherTarget(),
-            { _, event, _ -> OSStatus in
-                var hotKeyID = EventHotKeyID()
-                let status = GetEventParameter(
-                    event,
-                    EventParamName(kEventParamDirectObject),
-                    EventParamType(typeEventHotKeyID),
-                    nil,
-                    MemoryLayout<EventHotKeyID>.size,
-                    nil,
-                    &hotKeyID
-                )
-
-                if status == noErr {
-                    if let callback = HotKeyCallbacks.shared.callbacks[Int(hotKeyID.id)] {
-                        DispatchQueue.main.async {
-                            callback()
-                        }
-                    }
-                }
-
-                return noErr
-            },
-            1,
-            &eventType,
-            nil,
-            nil
-        )
-    }
-}
